Lots of things have changed. Haven't pushed in ages

+++ sh/colreplace.sh
@@ -3,23 +3,23 @@
 # a unique ID for the row, and a series of files of names and data to insert
 # into the specified columns
 
-if (($# && !(($#+1) % 3)))
+if (($# && !(($#-1) % 3)))
 then
     # file for original data table
     origfile=$1
     # number of columns used in the original data table for the field
     # which is being updated. 0 if the field is new
-    existingcols=$2
+#    existingcols=$2
 
     # make temporary files for storing intermediate results
     infile=$(mktemp)
     outfile=$(mktemp)
-
+    
     cp $origfile $outfile
     # remove heading from original file
 #    tail -n +2 $origfile > $outfile
     # count number of columns in original file
-    origcols=$(head -n 2 $origfile | tail -n 1 | wc -w)
+    origcols=$(head -n 1 $origfile | wc -w)
     # highest column that has been altered
     highestcol=0
     # amount of extra columns added so far
@@ -28,12 +28,12 @@ then
     # iterate over each set of 3 arguments, where the first is a file
     # of star names, the second is a file of data associated with those
     # star names, and the third is the first column to insert the data at
-    for (( i = 3; i <= $#; i += 3 ))
+    for (( i = 2; i <= $#; i += 3 ))
     do
-        eval "namefile=\$$i
-              datafile=\$$((i+1))
-              tablecol=\$$((i+2))"
-	if [ $tablecol -le $highestcol ]
+        eval "inputfile=\$$i
+              replaceindex=\$$((i+1))
+              replacecount=\$$((i+2))"
+	if [ $replaceindex -le $highestcol ]
 	then
 	    echo "Arguments must be given in ascending order."
 	    # delete tmp files
@@ -41,28 +41,29 @@ then
 	    exit 1
 	else
 	    # record the highest column used in the original table
-	    highestcol=$tablecol
+	    highestcol=$replaceindex
 	fi
-	# count columns in data file
-	datacols=$(head -n 1 $datafile | wc -w)
+	# count columns in data file to make sure there are enough
+	inputcols=$(head -n 1 $inputfile | wc -w)
+	if [ $inputcols -lt $replacecount ]
+	then
+	    echo \
+		"Number of columns in $inputfile is less than number requested"
+	    exit 1
+	fi
 	# copy output file to input file so that it can be used for output
 	cp $outfile $infile
 
-	# remove leading whitespace from namefile
-	sed "s/^[[:blank:]]*//" $namefile |
-	# now extract names, removing any file extensions
-	sed "s/[.[:blank:]].*//" |
-	# paste names with their data
-	paste - $datafile |
-	join $infile - > $outfile
+
+	join $infile $inputfile > $outfile
 	# output file becomes new input file
 	cp $outfile $infile
 	# now use awk to overwrite the appropriate columns, and save the
 	# result to the output file
-	awk -v existingcols=$existingcols \
+	awk -v existingcols=$replacecount \
             -v tablewidth=$((origcols+colshift)) \
-            -v startcol=$((tablecol+colshift)) \
-            -v newcols=$datacols '{
+            -v startcol=$((replaceindex+colshift+1)) \
+            -v newcols=$replacecount '{
         for (col = 1;                     col <  startcol;           col++)
             printf $col " ";
         for (col = tablewidth+1;          col <= tablewidth+newcols; col++)
@@ -72,7 +73,8 @@ then
         printf "\n"
         }' $infile > $outfile
 	# account for the change in the number of columns in the table
-	((colshift += datacols-existingcols))
+	colshift=0
+#	((colshift += replacecount))
     done
     # replace all whitespace with tabs
     # repeated whitespace is replaced with a single tab
@@ -82,8 +84,8 @@ then
     rm $infile $outfile
 else
     echo "\
-Usage: data-table0 [existing table] [existing PC #]
-                   ([name table] [data table] [column #])*
+Usage: colreplace [existing table]
+                  ([input table] [column to insert] [columns to insert])*
 
   column #'s must occur in ascending order."
 fi
